Unlink files when video delete

+++ app/Helpers/Helper.php
@@ -559,8 +559,12 @@
         }
 
         public static function delete_picture($picture, $path) {
+
+            if (file_exists(public_path() . $path . basename($picture))) {
             // "/uploads/"
-            File::delete( public_path() . $path . basename($picture));
+                File::delete( public_path() . $path . basename($picture));
+
+            }
             return true;
         }
